add support for --environment
This is automatically injected into the preprocessor evaluation
envValues for stack-args.yaml and can be used to conditionally load
different imports per environment. It is only in scope there and is
not intended for use anywhere beyond stack-args.yaml. It is NOT a
global value and is NOT propagated if the preprocessor is used on
cloudformation templates. Use normal CFN Parameters to pass
'environment' into those.

Additionally, Profile and Region can now be specified as maps keyed by
the environment name.

I've also snuck in some other minor tweaks to the cli args in this
commit.