Add request-approve-template command
the command is used as an approval process for the cfn-template.yaml.
When a template is uploaded to an s3 bucket through the cmd, it appends
the suffix `.pending` to the filename. This then enables the
maintainers/admins to review the changes in the cfn-template.yaml before
using it for the stack.

When a template has already been approved (the filename is not appended
with `.pending`) it results to a noop.

When a template is changed, the filename is changed because of the md5
hashing.