Forces iidy to wait for a terminal state when summarizing stack contents
For some reason, iidy is getting an in progress state when summarizing stack contents. It's possible this is just due to timing.
When we know we're going to be in a terminal state, we can just wait for a terminal status to be returned instead of accepting the in progress status.
It's not clear why this specifically happens with `UPDATE_COMPLETE_CLEANUP_IN_PROGRESS`, but it could be a special case for timing.